Create settings page and wire up the app router to navigate between board & settings.
